Add checkpatch.py
Peter recently posted a series to update the kernel-doc perl script with
the new python one from the upstream kernel tree. I mentioned on IRC I
had a personal/private checkpatch python script and was encouraged to
post this here on the list.

While there's nothing immediately wrong with checkpatch.pl, its design
makes it harder to implement new checks, refactor code, and because of
its custom nature it's been nigh impossible for years to sync it with
the kernel's original implementation. So there's no particular
attachment to staying with a perl script of many thousands of lines.

This patch introduces a script/checkpatch.py script that uses a
script/libcheckpatch python library that doesn't use any dependencies.

There's a proof of concept for unittests:

  python3 -m unittest -v scripts/libcheckpatch/tests.py

The CLI use is/should be identical to checkpatch.pl; just substitute the
filename extension in your invocations. The output format(s) are also
identical.

All checks that are in the original perl script and are missing are
annotated with a # TODO comment. A grep TODO | wc -l says it's about 60
at the time of writing.

This is an incomplete replacement for checkpatch.pl, so for this RFC I
look forward to high-level architecture and design comments.

NB: Python's import resolution logic makes it difficult to split checks for
different kinds of files into separate modules, unfortunately. I hit
"circular import" errors a lot.

# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

To: qemu-devel@nongnu.org
Cc: John Snow <jsnow@redhat.com>
Cc: Cleber Rosa <crosa@redhat.com>
Cc: Peter Maydell <peter.maydell@linaro.org>
Cc: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>
Cc: Richard Henderson <richard.henderson@linaro.org>
Cc: Philippe Mathieu-Daudé <philmd@linaro.org>
Cc: Alex Bennée <alex.bennee@linaro.org>
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: Mauro Carvalho Chehab <mchehab@kernel.org>
Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20250806-checkpatch-py-9ed21d18c99e",
    "prefixes": [
      "RFC WIP"
    ]
  }
}